Rename folder initializer => initializers
