+++ src/instance-initializers/logger.ts
@@ -1,5 +1,5 @@
 import type ApplicationInstance from '@ember/application/instance';
-import type { Logger } from '../initializer/logger';
+import type { Logger } from '../initializers/logger';
 
 export function initialize(applicationInstance: ApplicationInstance) {
   const logger = applicationInstance.lookup('logger:main') as Logger;
